refactor(vscode): lift history diff sub-views into historyDiffView.ts
Closes the history-side extraction started in #818 (pure helpers +
tests) and continued in #820 (timeline + row construction). Lifts the
remaining diff-rendering surface — the inline `.expanded.diff-expanded`
view that opens for "Diff vs previous" / "Diff vs current", plus the
auto-clearing two-way `.diff-inline` banner — out of `behavior.ts`
into a new sibling `historyDiffView.ts`.

After this PR `history/behavior.ts` is 253 lines of orchestration only:
DOM lookups, message dispatch, filter / scope plumbing, and the
initial render kickoff. Every imperative DOM operation now lives in a
typed sibling module.

New `historyDiffView.ts` exports:
- `fillDiff(id, against, leftLabel, leftImage, rightLabel, rightImage,
   config)` — populate the open expansion with the streamed diff
  payload. Builds the Side / Overlay / Onion mode bar (persisted via
  `vscode.setState`), kicks off async pixel-diff stats, renders the
  matching mode body.
- `computeDiffStats(left, right)` — async client-side pixel diff,
  resolves a typed `DiffStats` discriminated union (error / size
  mismatch / pixel count). Doesn't reject; every failure mode lands
  as an `{ error }` variant. Algorithm duplicated from
  `webview/preview/diffOverlay.ts` (deliberately, until a shared
  `pixelDiff.ts` collapses both copies).
- `applyDiffStats(el, s)` — render the stats label with `data-state`
  for CSS colouring (`identical` / `changed` / `size-mismatch`).
- `showDiff(fromId, toId, result, config)` — toolbar-driven inline
  banner for the shift-click two-way diff path; auto-clears after 12s.
- `DiffStats` type union exported for tests / future reuse.

`HistoryDiffViewConfig` is the slim collaborator surface — `vscode`
(typed `VsCodeApi<unknown>`; `fillDiff` casts the persisted state to
`PersistedHistoryState` on read so callers don't have to thread the
type parameter through) and `timelineEl`. `behavior.ts` builds the
config once at panel boot and threads it through the four
`diffResult` / `diffError` / `diffReady` message handlers.

## Drive-by

Drops three now-unused imports from `behavior.ts`:
- `buildDiffModeBar` and `DiffMode` from `../shared/diffModeBar` —
  only consumed inside the lifted `fillDiff`.
- `HistoryDiffSummary` from `../shared/types` — `showDiff` was the
  sole consumer.

`history/behavior.ts`: 579 → 253 lines (−326 LOC).
`historyDiffView.ts`: +379 lines (new file).

`history/behavior.ts` is now down 66% from its 745-line `@ts-nocheck`
starting point at the start of this round, and 92% from its post-#767
peak. The history panel mirrors the preview panel's split: a thin
`behavior.ts` orchestration layer over a dozen typed sibling modules.

## Test plan

- [x] `npm run compile` clean (host + webview)
- [x] `npm test` — 373/373 passing
- [x] `npm run format:check` clean
- [ ] Manual smoke:
  - "Diff vs previous" / "Diff vs current" per-row buttons open an
    inline diff expansion; cycle Side / Overlay / Onion modes; mode
    persists across panel reload via `vscode.setState`.
  - Onion-skin slider mixes left/right base/top images.
  - Pixel-diff stats label renders identical / changed / size-mismatch
    states with the right `data-state`.
  - Shift+click two rows + toolbar Diff posts a `diffResult` /
    `diffError`; the inline banner appears at the top of the timeline
    and auto-clears after 12s.
